`zcashd-regtest-controller: Move setup/teardown into submods; use `zcash.conf` rather than cli args.
