+++ Cargo.toml
@@ -16,6 +16,7 @@ anyhow = "1.0.75"
 anyhow-std = "0.1.4"
 async-trait = "0.1.74"
 derive_more = "0.99.17"
+indoc = "2.0.4"
 rand = "0.8.5"
 rand_chacha = "0.3.1"
 reqwest = { version = "0.11.22", features = ["json"] }
@@ -24,6 +25,6 @@ serde = { version = "1.0.189", features = ["derive"] }
 tempfile = "3.8.0"
 thiserror = "1.0.50"
 tokio = { version = "1.33.0", features = ["rt", "macros"] }
+zcashd-regtest-controller = { path = "./regtest-controller" }
 zcashd-rpc-messages = { path = "./messages" }
 zcashd-rpc-provider = { path = "./provider" }
-zcashd-regtest-controller = { path = "./regtest-controller" }
