+++ regtest-controller/Cargo.toml
@@ -6,5 +6,6 @@ edition.workspace = true
 [dependencies]
 anyhow.workspace = true
 anyhow-std.workspace = true
+indoc.workspace = true
 tempfile.workspace = true
 tokio = { workspace = true, features = ["fs", "process"] }
